# bunch queries for content types