--- a/opal/core/api.py
+++ b/opal/core/api.py
@@ -5,12 +5,9 @@ import collections
 
 from django.conf import settings
 from django.views.generic import View
+from django.contrib.contenttypes.models import ContentType
 from rest_framework import routers, status, viewsets
-from rest_framework.decorators import list_route
-from rest_framework.views import APIView
 from rest_framework.response import Response
-from rest_framework.reverse import reverse
-
 from opal.models import Episode
 from opal.core import application, exceptions, plugins
 from opal.core import glossolalia
@@ -121,19 +118,21 @@ class OptionsViewSet(viewsets.ViewSet):
         from opal.models import Synonym, Team, Macro
 
         data = {}
-        for model in LookupList.__subclasses__():
+        subclasses = LookupList.__subclasses__()
+        for model in subclasses:
+            options = list(model.objects.all().values_list("name", flat=True))
             options = [instance.name for instance in model.objects.all()]
             data[model.__name__.lower()] = options
 
-        for synonym in Synonym.objects.all():
-            try:
-                co =  synonym.content_object
-            except AttributeError:
-                continue
-            if co is None:
-                continue
-            name = type(co).__name__.lower()
-            data[name].append(synonym.name)
+        model_to_ct = ContentType.objects.get_for_models(
+            *subclasses
+        )
+
+        for model, ct in model_to_ct.iteritems():
+            synonyms = Synonym.objects.filter(content_type=ct).values_list(
+                "name", flat=True
+            )
+            data[model.__name__.lower()].extend(synonyms)
 
         for name in data:
             data[name].sort()