local speed up

old
super GET /api/v0.1/options/ 200 (1.99 seconds) (365 SQL queries, 1869.0 ms)

new
super GET /api/v0.1/options/ 200 (1.13 seconds) (214 SQL queries, 1489.0 ms)