@davidmiller 2015-11-30
@fredkingham Do we have tests coverage for this API ?